Adds self-discovery of the GitHub site url.
Looks at remote.origin.url to make it's determination to extract the host.  If there isn't a remote.origin.url, then NoRemoteRepository is raised.

if remote.origin.url is a "normal" URI, then the host is extracted normally.  If it's an SSH URI starting with 'git@' (which is "standard" git), then it's extracted using a regex.

If the host is in the github.com domain, then the site is set to https://api.github.com.  Anything else and the site it "http://#{host}".