Adds git to-master {repoName}/{branchNumber}
If a number is given to "git-to-master", or a number with a server
preceding it (e.g., "origin/23"), then this assumes that the number
refers to an existing pull request identified by that number. In
that case, the branch associated with the HEAD of the pull request
is checked out before doing the rest of the "to-master".

Closes GH-94